WIP: Disable scanning of printers to prevent accidental resets.
This disables the discovery of which serial ports are printers, because
when connecting, it causes printers to reset, which can interrupt a
print that was started from outside this cura instance.
For now, we only add a single USBPrinter device with 'Autodetect' id
as long as we find available serial devices.