Fix "Wrong Z on successive prints w/o power reset" (T1466)
The error in Z was caused by the vertical move immediately after G28.
On successive prints, the bed compensation was adjusting this
vertical path and introducing lateral motion, potentially causing it
to trigger the X or Y endstop.

The fix modified the start GCODE to move away from the endstops
right after G28.